requirements.txt
fastapi>=0.124
uvicorn[standard]>=0.30
pyrtlsdr==0.3.0
numpy>=2.0
scipy>=1.15
anthropic>=0.86
twilio>=9.10
python-dotenv>=1.0
httpx>=0.27
faster-whisper>=1.0
